- Kernel paging working! - Add a disassembly dump of boot sector with symbols and source - Use INT 13h extensions to read disk in bootloader - Updated to use multi-sector reads - Reduce alignment in bootloader linker script to save space - Remove code16gcc.h headers and instead use -m16 compiler option - Remove unnecessary stack arg in entry to kernel - Map the whole 640KB range before entry to kernel (probably not needed now though) - Print the entry point in bootloader, before entering kernel - Add a way to map an address range without allocation semantics in bootloader - Add -fomit-frame-pointer in kernel - Moved mmu_set_fsgsbase - Renamed msr.c to more general control_regs.c - Added functions to get/set the page directory register and invalidate TLB entries - Added table driven flag bit formatter - Added descriptive text about eflags and mxcsr in crash dump - Added console wrappers for console output calls - Fixed bug in __dvec2 typedef - Made console cursor_toggle return the old shown state - Made everything depend upon Makefile in makefile - Fixed -c handling in autodep.mk - Added processor agnostic thread.h - vcprintf disables the cursor and reenables it later for efficiency - printk now returns the number of chars written - Added aligned16_memset and optimized memset to use SSE 128-bit copies - Added hang avoidance in sleep for when timer isn't initialized yet - Use variables for most programs in Makefile - Comment out large block of test hack code in main

+++ kernel/arch/x86_64/types.h
@@ -38,7 +38,7 @@ typedef uint16_t __uvec8 __attribute__((vector_size(16)));
 typedef uint32_t __uvec4 __attribute__((vector_size(16)));
 typedef uint64_t __uvec2 __attribute__((vector_size(16)));
 typedef float __fvec4 __attribute__((vector_size(16)));
-typedef float __dvec2 __attribute__((vector_size(16)));
+typedef double __dvec2 __attribute__((vector_size(16)));
 
 #define countof(arr) (sizeof(arr)/sizeof(*arr))
